requirements.txt
flask
flask_restful
requests
redis
sqlalchemy
psycopg2
python-dotenv